requirements.txt
langchain
langchain_community
langchain_core
langchain_ollama
langchain_openai
langgraph
ollama==0.4.7
typing_extensions
fastapi
uvicorn
pydantic
python-dotenv